Consolidated models into one place
I removed the models folder from the CronUpdater. And now CronUpdater
references the models in app_api instead.
