--- CronUpdater/models/activeAlerts/activePost.js
@@ -1,11 +1,0 @@
-var mongoose = require('mongoose');
-
-var activePostSchema = mongoose.Schema({
-    agency: String,
-    imageLink: String,
-    title: String,
-    postInfo: String,
-    time: Date
-});
-
-mongoose.model('ActivePost', activePostSchema);
