--- CronUpdater/models/activeAlerts/activeWeatherPost.js
@@ -1,11 +1,0 @@
-var mongoose = require('mongoose');
-
-var activeWeatherPostSchema = mongoose.Schema({
-    agency: String,
-    imageLink: String,
-    title: String,
-    postInfo: String,
-    time: Date
-});
-
-mongoose.model('ActiveWeatherPost', activeWeatherPostSchema);
