--- CronUpdater/models/loggedAlerts/post.js
@@ -1,11 +1,0 @@
-var mongoose = require('mongoose');
-
-var postSchema = mongoose.Schema({
-    agency: String,
-    imageLink: String,
-    title: String,
-    postInfo: String,
-    time: Date
-});
-
-mongoose.model('Post', postSchema);
