--- CronUpdater/models/loggedAlerts/wAlert.js
@@ -1,8 +1,0 @@
-var mongoose = require('mongoose');
-
-var wAlertSchema = mongoose.Schema({
-    AlertID: String,
-    AlertInfo: String
-});
-
-mongoose.model('wAlert', wAlertSchema, 'wAlerts');
